Combined ImageKeeper and ToolController parameters to UndoableCommand into one UndoContext interface.
By doing this all the information needed to undo/redo a command are centralized in one place. Also the getSelectedTool method of the ToolController interface was never used in any command and thus the interface was actually too broad. This change will also allow to add new things that become necessary to undo/redo something to the interface.